rename files to be consistent with class names
